+++ packages/sdrr/CHANGELOG.md
@@ -1,5 +1,12 @@
 # soda-node
 
+## 0.1.14
+
+### Patch Changes
+
+- Updated dependencies
+    - deepsea-tools@5.47.5
+
 ## 0.1.13
 
 ### Patch Changes